added default roles and user to database when database is created

+++ src/main/java/com/hazr/personalblog/model/Role.java
@@ -10,7 +10,7 @@ public class Role implements GrantedAuthority {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "role_id")
-    private Integer roleId;
+    private Long roleId;
 
     @Column(name = "authority")
     private String authority;
@@ -25,7 +25,7 @@ public class Role implements GrantedAuthority {
         this.authority = authority;
     }
 
-    public Role(Integer roleId, String authority) {
+    public Role(Long roleId, String authority) {
         this.roleId = roleId;
         this.authority = authority;
     }
@@ -40,11 +40,11 @@ public class Role implements GrantedAuthority {
         this.authority = authority;
     }
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return this.roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
